--- src/NServiceBus.Core/Recoverability/Recoverability.cs
@@ -1,31 +1,0 @@
-﻿namespace NServiceBus
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using DelayedDelivery;
-    using DeliveryConstraints;
-    using Faults;
-    using Features;
-    using Hosting;
-    using Logging;
-    using Settings;
-    using Support;
-    using Transport;
-
-    class Recoverability : Feature
-    {
-        public Recoverability()
-        {
-            EnableByDefault();
-            DependsOnOptionally<DelayedDeliveryFeature>();
-
-            Prerequisite(context => !context.Settings.GetOrDefault<bool>("Endpoint.SendOnly"),
-                "Message recoverability is only relevant for endpoints receiving messages.");
-        }
-
-        protected internal override void Setup(FeatureConfigurationContext context)
-        {
-        }
-    }
-}
